% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irm.R
\name{bootstrap_irm}
\alias{bootstrap_irm}
\title{Bootstrap Implementation for Interactive Regression Model}
\usage{
bootstrap_irm(theta, g0_hat, g1_hat, u0_hat, u1_hat, d, m, inf_model, se,
  bootstrap, nRep)
}
\arguments{
\item{theta}{Final dml estimator for the interactive regression model.}

\item{u0_hat}{Residuals from \eqn{y-g(0,x)}.}

\item{u1_hat}{Residuals from \eqn{y-g(1,x)}.}

\item{d}{Treatment variable.}

\item{m}{Predictions from \eqn{d-m(x)}.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"ATE"}. Alternatively switch to \code{"ATET"} for average treatment effect on the treated.}

\item{se}{Estimated standard error from DML procedure.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}
}
\value{
List with bootstrapped standard errors (\code{boot_se}) and bootstrapped coefficients.
}
\description{
Multiplier bootstrap to construct simultaneous confidence bands for multiple target coefficients in a interactive regression model (irm) with double machine learning.
}
