% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irmiv.R
\name{bootstrap_irmiv}
\alias{bootstrap_irmiv}
\title{Bootstrap Implementation for Interactive Instrumental Variable Regression Model}
\usage{
bootstrap_irmiv(theta, p_hat, mu0_hat, mu1_hat, m0_hat, m1_hat, d, y, z,
  inf_model, se, bootstrap, nRep)
}
\arguments{
\item{theta}{Final dml estimator for interactive regression model.}

\item{d}{Treatment variable.}

\item{y}{Outcome variable.}

\item{z}{Instrumental variable.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"LATE"}. Alternatively switch to \code{"LATET"} for average treatment effect on the treated.}

\item{se}{Estimated standard error from DML procedure.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}
}
\value{
List with bootstrapped standard errors (\code{boot_se}) and bootstrapped coefficients.
}
\description{
Multiplier bootstrap to construct simultaneous confidence bands for multiple target coefficients in a interactive regression model (irm) with double machine learning.
}
