% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plriv.R
\name{bootstrap_plriv}
\alias{bootstrap_plriv}
\title{Bootstrap Implementation for Partially Linear Regression Model}
\usage{
bootstrap_plriv(theta, u_hat, v_hat, w_hat, inf_model, se, bootstrap, nRep)
}
\arguments{
\item{theta}{final dml estimator for the partially linear model.}

\item{u_hat}{Residuals from \eqn{y-g(x)}.}

\item{v_hat}{Residuals from \eqn{d-m(x)}.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{se}{Estimated standard error from DML procedure.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}
}
\value{
List with bootstrapped standard errors (\code{boot_se}) and bootstrapped coefficients.
}
\description{
Multiplier bootstrap to construct simultaneous confidence bands for multiple target coefficients in a partially linear regression model (PLR) with double machine learning.
}
