% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{coef.DML}
\alias{coef.DML}
\title{Methods for Inference Task}
\usage{
\method{coef}{DML}(object, ...)
}
\arguments{
\item{object}{an object of class \code{DML}.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Methods for S3 class \code{DML}
}
