% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{coef.InfTask}
\alias{coef.InfTask}
\title{Methods for Inference Task}
\usage{
\method{coef}{InfTask}(object, ...)
}
\arguments{
\item{object}{an object of class \code{InfTask}.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Methods for S3 class \code{InfTask}
}
