% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{confint.DML}
\alias{confint.DML}
\alias{summary.DML}
\title{Methods for Inference Task}
\usage{
\method{confint}{DML}(object, parm, level = 0.95, joint = FALSE, ...)

\method{summary}{DML}(object, ...)
}
\arguments{
\item{object}{Object of class \code{DML}.}

\item{parm}{a specification of which parameters are to be given confidence intervals among the variables for which inference was done, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}

\item{joint}{logical, if \code{TRUE} joint confidence intervals are calculated.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Methods for S3 class \code{DML}

Summary method for class \code{DML}.
}
