% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{DML}
\alias{DML}
\title{Wrapper Function for Double Machine Learning with Cross-Fitting}
\usage{
DML(data, y, d, z = NULL, model = "plr", k = 2, S = 1,
  resampling = NULL, ResampleInstance = NULL, mlmethod,
  dml_procedure = "dml2", params = list(params_m = list(), params_g =
  list()), inf_model = "IV-type", se_type = "ls",
  bootstrap = "normal", nRep = 500, aggreg_median = TRUE, ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{z}{Name of instrument variable.}

\item{model}{Inference model to be implemented, e.g. partially linear regression (\code{plr}) (default).}

\item{k}{Number of folds for \code{k}-fold cross-fitting (default \code{k}=2).
#' @param S number of repetitions (default \code{S=1}.}

\item{resampling}{Resampling scheme for cross-fitting of class \code{"ResampleDesc"}.}

\item{ResampleInstance}{(Optional) \code{ResampleInstance} that can be passed through in order to obtain replicable sample splits. By default, \code{ResampleInstance} is set \code{NULL} and \code{resampling} is instantiated internally. Note that \code{ResampleInstance} will override the information in \code{resampling}.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_g} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X (nuisance part g). Set \code{mlmethod_m} for  classification or regression method for regression of d on X (nuisance part m). If multiple target coefficients are provided and different mlmethods chosen for each coefficient, the names of the methods are required to match. A list of available methods is available at \url{https://mlr.mlr-org.com/articles/tutorial/integrated_learners.html}.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_g} for predictions of nuisance part g and \code{params_m} for nuisance m. If multiple target coefficients are provided the names of the lists with hyperparameters and the target coefficients must match.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{se_type}{Method to estimate standard errors. Default \code{"ls"} to estimate usual standard error from least squares regression of residuals. Alternatively, specify \code{"IV-type"} or \code{"DML2018"} to obtain standard errors that correspond to the specified \code{inf_model}. The options chosen for \code{inf_model} and \code{se_type} are required to match.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}

\item{aggreg_median}{logical indicating how target estimators and standard errors should be aggregated for repeated cross-fitting. By default (\code{TRUE}), the median over all estimators as taken as the final estimator. If \code{FALSE}, the mean is calculated over all estimators.}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object of class \code{DML} with estimated coefficient and standard errors.
}
\description{
Implements double machine learning inference with \code{k}-fold cross-fitting (default).
}
