% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irm.R
\name{dml_irm}
\alias{dml_irm}
\title{Double Machine Learning for Interactive Regression Model.}
\usage{
dml_irm(data, y, d, k = 2, resampling = NULL, mlmethod,
  params = list(params_m = list(), params_g1 = list(), params_g2 =
  list()), dml_procedure = "dml2", inf_model = "ATE",
  se_type = "ATE", bootstrap = "normal", nRep = 500, ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{k}{Number of folds for \code{k}-fold cross-fitting (default \code{k}=2).}

\item{resampling}{Resampling scheme for cross-fitting of class \code{\link[mlr3]{ResamplingCV}}.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_g} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X (nuisance part g). Set \code{mlmethod_m} for  classification or regression method for regression of d on X (nuisance part m).}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_g1} for predictions of nuisance part g1, \code{params_g0} for nuisance part g0, and \code{params_m} for nuisance m.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"ATE"}. Alternatively switch to \code{"ATET"} for average treatment effect on the treated.}

\item{se_type}{Method to estimate standard errors (redundant / identical to inf_model).}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object with estimated coefficient and standard errors.
}
\description{
Double Machine Learning for Interactive Regression Model.
}
