% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irmiv.R
\name{dml_irmiv}
\alias{dml_irmiv}
\title{Double Machine Learning for Interactive Instrumental Variable Regression Model.}
\usage{
dml_irmiv(data, y, d, z, k = 2, resampling = NULL, mlmethod,
  params = list(params_mu0 = list(), params_mu1 = list(), params_m0 =
  list(), params_m1 = list(), params_p = list()), dml_procedure = "dml2",
  always_takers = TRUE, never_takers = TRUE, inf_model = "LATE",
  se_type = "LATE", bootstrap = "normal", nRep = 500, ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{z}{Name of instrumental variables.}

\item{k}{Number of folds for \code{k}-fold cross-fitting (default \code{k}=2).}

\item{resampling}{Resampling scheme for cross-fitting of class \code{\link[mlr3]{ResamplingCV}}.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_mu1} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X for individuals with Z = 1  (nuisance part mu1) and \code{mlmethod_mu0} for individuals with Z = 0 (nuisance part mu0). Set \code{mlmethod_m1} for  classification method for regression of d on X for individuals with Z = 1 (nuisance part m1) and \code{mlmethod_m0} for observations with Z = 0 (nuisance part m0). Set \code{mlmethod_p} for classification method for propensity score (nuisance part p).}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_mu1} for predictions of nuisance part g1, \code{params_mu0} for nuisance part mu0, and so forth (names must match those in \code{mlmethod}).}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{always_takers}{option to adapt to cases with (default) and without always-takers. If \code{FALSE}, the estimator is adapted to a setting without always-takers.}

\item{never_takers}{option to adapt to cases with (default) and without never-takers. If \code{FALSE}, the estimator is adapted to a setting without never-takers.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"LATE"}. Alternatively switch to \code{"LATET"} for average treatment effect on the treated.}

\item{se_type}{Method to estimate standard errors (redundant / identical to inf_model).}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object with estimated coefficient and standard errors.
}
\description{
Double Machine Learning for Interactive Instrumental Variable Regression Model.
}
