% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plr.R
\name{dml_plr}
\alias{dml_plr}
\title{Double Machine Learning for Partially Linear Regression.}
\usage{
dml_plr(data, y, d, resampling, mlmethod, params = list(params_m =
  list(), params_g = list()), dml_procedure = "dml2",
  inf_model = "DML2018", ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{resampling}{Resampling scheme for cross-fitting of class \code{"ResampleDesc"}.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_g} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X (nuisance part g). Set \code{mlmethod_m} for  classification or regression method for regression of d on X (nuisance part m). A list of available methods is available at \url{https://mlr.mlr-org.com/articles/tutorial/integrated_learners.html}.}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_g} for predictions of nuisance part g and \code{params_m} for nuisance m.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object with estimated coefficient and standard errors.
}
\description{
Double Machine Learning for Partially Linear Regression.
}
