% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plriv.R
\name{dml_plriv}
\alias{dml_plriv}
\title{Double Machine Learning for Partially Linear Regression.}
\usage{
dml_plriv(data, y, d, z, k = 2, resampling = NULL, mlmethod,
  params = list(params_m = list(), params_r = list(), params_g = list()),
  dml_procedure = "dml2", inf_model = "partialling-out",
  se_type = "partialling-out", bootstrap = "normal", nRep = 500, ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{z}{Name of instrumental variables.}

\item{k}{Number of folds for \code{k}-fold cross-fitting (default \code{k}=2).}

\item{resampling}{Resampling scheme for cross-fitting of class \code{\link[mlr3]{ResamplingCV}}.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_g} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X (nuisance part g). Set \code{mlmethod_m} for  classification or regression method for regression of z on X (nuisance part m). Set \code{mlmethod_m} for  classification or regression method for regression of d on X (nuisance part r).}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_g} for predictions of nuisance part g, \code{params_m} for nuisance m, and \code{params_r} for nuisance r.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{inf_model}{Inference model for final estimation, default \code{"partialling-out"} (...)}

\item{se_type}{Method to estimate standard errors. Default \code{"partialling-out"}. The options chosen for \code{inf_model} and \code{se_type} are required to match.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object with estimated coefficient and standard errors.
}
\description{
Double Machine Learning for Partially Linear Regression.
}
