% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_ml_data.R
\docType{data}
\name{DoubleMLData}
\alias{DoubleMLData}
\title{DoubleMLData}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
DoubleMLData
}
\description{
Data structure for double machine learning.
}
\keyword{datasets}
