% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_ml_iivm.R
\docType{class}
\name{DoubleMLIIVM}
\alias{DoubleMLIIVM}
\title{DoubleMLIIVM R6 class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DoubleMLIIVM
}
\description{
DoubleMLIIVM R6 class
}
\keyword{datasets}
