% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_ml_pliv.R
\docType{class}
\name{DoubleMLPLIV}
\alias{DoubleMLPLIV}
\title{DoubleMLPLIV R6 class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
DoubleMLPLIV
}
\description{
DoubleMLPLIV R6 class
}
\keyword{datasets}
