% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_ml_plr.R
\docType{class}
\name{DoubleMLPLR}
\alias{DoubleMLPLR}
\title{DoubleMLPLR R6 class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DoubleMLPLR
}
\description{
DoubleMLPLR R6 class
}
\keyword{datasets}
