% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helpers}
\alias{helpers}
\alias{extract_test_pred}
\alias{format.perc}
\alias{se_repeated}
\title{Helper Functions}
\usage{
extract_test_pred(x)

\method{format}{perc}(probs, digits)

se_repeated(se_s, coefficients, theta_s, aggreg_median)
}
\arguments{
\item{x}{some object.}

\item{probs}{probability.}

\item{digits}{number of digits.}

\item{se_s}{Standard errors from repeated cross-fitting.}

\item{coefficients}{Final estimators aggregated over all repetitions.}

\item{theta_s}{Estimators from all repetitions.}

\item{aggreg_median}{logical indicating how target estimators and standard errors should be aggregated for repeated cross-fitting. By default (\code{TRUE}), the median over all estimators as taken as the final estimator. If \code{FALSE}, the mean is calculated over all estimators.}
}
\description{
Functions needed in package code.
}
