% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plr.R
\name{InferenceTask}
\alias{InferenceTask}
\title{Wrapper Function for Double Machine Learning Inference}
\usage{
InferenceTask(data, y, d, z = NULL, model = "plr", k = 2,
  resampling = NULL, mlmethod, dml_procedure = "dml2",
  params = list(), inf_model = "DML2018", ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{z}{Name of instrument variable.}

\item{model}{Inference model to be implemented, e.g. partially linear regression (\code{plr}) (default).}

\item{k}{Number of folds for \code{k}-fold cross-fitting (default \code{k}=2).}

\item{resampling}{Resampling scheme for cross-fitting of class \code{"ResampleDesc"}, default 2-fold cross-fitting.}

\item{mlmethod}{List with classification or regression methods according to naming convention of the \code{mlr} package. Set \code{mlmethod_g} for classification or regression method according to naming convention of the \code{mlr} package for regression of y on X (nuisance part g). Set \code{mlmethod_m} for  classification or regression method for regression of d on X (nuisance part m). A list of available methods is available at \url{https://mlr.mlr-org.com/articles/tutorial/integrated_learners.html}.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{params}{Hyperparameters to be passed to classification or regression method. Set hyperparameters \code{params_g} for predictions of nuisance part g and \code{params_m} for nuisance m.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object of class \code{InfTask} with estimated coefficient and standard errors.
}
\description{
Implements double machine learning inference with \code{k}-fold cross-fitting (default).
}
