% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irm.R
\name{orth_irm_dml}
\alias{orth_irm_dml}
\title{Orthogonalized Estimation of Coefficient in irm}
\usage{
orth_irm_dml(g0_hat, g1_hat, u0_hat, u1_hat, d, m, inf_model)
}
\arguments{
\item{g0_hat}{Predictions from \eqn{g(0,X)}.}

\item{g1_hat}{Predictions from \eqn{g(1,X)}}

\item{u0_hat}{Residuals from \eqn{y-g(0,x)}.}

\item{u1_hat}{Residuals from \eqn{y-g(1,x)}.}

\item{d}{Treatment variable.}

\item{m}{Predictions from \eqn{d-m(x)}.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"ATE"}. Alternatively switch to \code{"ATET"} for average treatment effect on the treated.}
}
\value{
List with estimate (\code{theta}).
}
\description{
Function to estimate the structural parameter in an interactive regression model (irm).
}
