% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plr.R
\name{orth_plr_dml}
\alias{orth_plr_dml}
\title{Orthogonalized Estimation of Coefficient in PLR}
\usage{
orth_plr_dml(u_hat, v_hat, v_hatd, inf_model)
}
\arguments{
\item{u_hat}{Residuals from \eqn{y-g(x)}.}

\item{v_hat}{Residuals from \eqn{d-m(x)}.}

\item{v_hatd}{Product \code{v_hat * d}.}

\item{inf_model}{Method to estimate structural parameter.}
}
\value{
List with estimate (\code{theta}).
}
\description{
Function to estimate the structural parameter in a partially linear regression model (PLR).
}
