% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plriv.R
\name{orth_plriv_dml}
\alias{orth_plriv_dml}
\title{Orthogonalized Estimation of Coefficient in PLR}
\usage{
orth_plriv_dml(u_hat, v_hat, w_hat, inf_model)
}
\arguments{
\item{u_hat}{Residuals from \eqn{y-g(x)}.}

\item{v_hat}{Residuals from \eqn{d-m(x)}.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}
}
\value{
List with estimate (\code{theta}).
}
\description{
Function to estimate the structural parameter in a partially linear regression model with instrumental variables (PLRIV).
}
