% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust}
\alias{p_adjust}
\alias{p_adjust.InfTask}
\title{Multiple Testing Adjustment of p-values for S3 objects \code{InfTask}}
\usage{
p_adjust(x, ...)

\method{p_adjust}{InfTask}(x, method = "RW", ...)
}
\arguments{
\item{x}{an object of S3 class \code{InfTask}.}

\item{...}{further arguments passed on to methods.}

\item{method}{the method of p-value adjustment for multiple testing.
Romano-Wolf stepdown ('\code{RW}') is chosen by default.}
}
\value{
A matrix with the estimated coefficients and the p-values that are
 adjusted according to the specified method.
}
\description{
Multiple hypotheses testing adjustment of p-values for double machine learning.
}
\details{
Multiple testing adjustment is performed for S3 objects of class
\code{InfTask}. Implemented methods for multiple testing
adjustment are Romano-Wolf stepdown '\code{RW}' (default) and the adjustment
methods available in the \code{p.adjust} function of the \code{stats} package,
including the Bonferroni, Bonferroni-Holm, and Benjamini-Hochberg corrections,
see \code{\link{p.adjust.methods}}.

Objects of class \code{InfTask} are constructed by
\code{\link{InferenceTask}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{InfTask}: 
}}

\references{
J.P. Romano, M. Wolf (2005). Exact and approximate stepdown
 methods for multiple hypothesis testing. Journal of the American Statistical
 Association, 100(469), 94-108.

J.P. Romano, M. Wolf (2016). Efficient computation of adjusted
 p-values for resampling-based stepdown multiple testing. Statistics and
 Probability Letters, (113), 38-40.
}
