% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{print.dca}
\alias{print.dca}
\title{Print an dca object}
\usage{
\method{print}{dca}(x, ...)
}
\arguments{
\item{x}{The result from dca function, which contains a list of the DCA results:
\item{B}{DCA suggested Mahalanobis matrix}
\item{DCA}{DCA suggested transformation of the data.
           The dimension is (original data dimension) * (useD)}
\item{newData}{DCA transformed data}}

\item{...}{ignored}
}
\description{
Print an dca object
}

