% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmd.r
\name{print.GdmDiag}
\alias{print.GdmDiag}
\title{Print an GdmDiag object}
\usage{
\method{print}{GdmDiag}(x, ...)
}
\arguments{
\item{x}{The result from dca function, which contains a list of the GdmDiag results:
\item{newData}{GdmDiag transformed data}
\item{diagonalA}{suggested Mahalanobis matrix}
\item{dmlA}{matrix to transform data, square root of diagonalA }
\item{error}{the precision of obtained distance metric by Newton-Raphson optimization }}

\item{...}{ignored}
}
\description{
Print an GdmDiag object
}

