% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmf.r
\name{print.gmdf}
\alias{print.gmdf}
\title{Print a gmdf object}
\usage{
\method{print}{gmdf}(x, ...)
}
\arguments{
\item{x}{The result from gmdf function, which contains TK}

\item{...}{ignored}
}
\description{
Print a gmdf object
}

