% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca.R
\name{print.rca}
\alias{print.rca}
\title{Print an rca object}
\usage{
\method{print}{rca}(x, ...)
}
\arguments{
\item{x}{The result from rca function, which contains mahalanobis metric,
whitening transformation matrix, and transformed data}

\item{...}{ignored}
}
\description{
Print an rca object
}

