% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{print.summary.DML}
\alias{print.summary.DML}
\title{Summarizing Inference Task}
\usage{
\method{print}{summary.DML}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{summary.DML}, usually a result of a call or \code{summary.DML}}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Summary method for class \code{DML}.
}
