% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{print.summary.InfTask}
\alias{print.summary.InfTask}
\title{Summarizing Inference Task}
\usage{
\method{print}{summary.InfTask}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{summary.InfTask}, usually a result of a call or \code{summary.InfTask}}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Summary method for class \code{InfTask}.
}
