% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_irmiv.R
\name{var_irmiv}
\alias{var_irmiv}
\title{Variance estimation for DML estimator in the Interactive Instrumental Variable Regression Model}
\usage{
var_irmiv(theta, p_hat, mu0_hat, mu1_hat, m0_hat, m1_hat, d, y, z,
  inf_model)
}
\arguments{
\item{theta}{Final dml estimator for interactive regression model.}

\item{d}{Treatment variable.}

\item{y}{Outcome variable.}

\item{z}{Instrumental variable.}

\item{inf_model}{Estimator for final estimation, default average treatment effect \code{"LATE"}. Alternatively switch to \code{"LATET"} for average treatment effect on the treated.}

\item{m}{Predictions from \eqn{d-m(x)}.}

\item{g0_hat}{Predictions from \eqn{g(0,X)}.}

\item{g1_hat}{Predictions from \eqn{g(1,X)}}

\item{u0_hat}{Residuals from \eqn{y-g(0,x)}.}

\item{u1_hat}{Residuals from \eqn{y-g(1,x)}.}
}
\value{
Variance estimator (\code{var}).
}
\description{
Variance estimation for the structural parameter estimator in an interactive regression model (irm) with double machine learning.
}
