% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plr.R
\name{var_plr}
\alias{var_plr}
\title{Variance estimation for DML estimator in the partially linear regression model}
\usage{
var_plr(theta, d, u_hat, v_hat, v_hatd, inf_model, se_type)
}
\arguments{
\item{theta}{final dml estimator for the partially linear model.}

\item{d}{treatment variable.}

\item{u_hat}{Residuals from \eqn{y-g(x)}.}

\item{v_hat}{Residuals from \eqn{d-m(x)}.}

\item{v_hatd}{Product of \code{v_hat} with \code{d}.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{se_type}{Method to estimate standard errors. Default \code{"ls"} to estimate usual standard error from least squares regression of residuals. Alternatively, specify \code{"IV-type"} or \code{"DML2018"} to obtain standard errors that correspond to the specified \code{inf_model}. The options chosen for \code{inf_model} and \code{se_type} are required to match.}
}
\value{
Variance estimator (\code{var}).
}
\description{
Variance estimation for the structural parameter estimator in a partially linear regression model (PLR) with double machine learning.
}
