% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plriv.R
\name{var_plriv}
\alias{var_plriv}
\title{Variance estimation for DML estimator in the partially linear regression model}
\usage{
var_plriv(theta, u_hat, v_hat, w_hat, inf_model, dml_procedure)
}
\arguments{
\item{theta}{final dml estimator for the partially linear model.}

\item{u_hat}{Residuals from \eqn{y-g(x)}.}

\item{v_hat}{Residuals from \eqn{d-r(x)}.}

\item{w_hat}{Residuals from \eqn{z-m(x)}.}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{d}{treatment variable.}
}
\value{
Variance estimator (\code{var}).
}
\description{
Variance estimation for the structural parameter estimator in a partially linear regression model (PLR) with double machine learning.
}
