% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{coef.InfTask}
\alias{coef.InfTask}
\title{Methods for Inference Task}
\usage{
\method{coef}{InfTask}(x)
}
\arguments{
\item{x}{Object of class \code{InfTask}.}
}
\description{
Methods for S3 class \code{InfTask}
}
