% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_plr.R
\name{dml_plr}
\alias{dml_plr}
\title{Double Machine Learning for Partially Linear Regression.}
\usage{
dml_plr(data, y, d, resampling = NULL, ResampleInstance = NULL,
  mlmethod, params = list(params_m = list(), params_g = list()),
  dml_procedure = "dml2", inf_model = "IV-type", se_type = "ls",
  bootstrap = "normal", nRep = 500, ...)
}
\arguments{
\item{data}{Data frame.}

\item{y}{Name of outcome variable. The variable must be included in \code{data}.}

\item{d}{Name of treatment variables for which inference should be performed.}

\item{resampling}{Resampling scheme for cross-fitting of class \code{"ResampleDesc"}.}

\item{ResampleInstance}{(Optional) \code{ResampleInstance} that can be passed through in order to obtain replicable sample splits. By default, \code{ResampleInstance} is set \code{NULL} and \code{resampling} is instantiated internally. Note that \code{ResampleInstance} will override the information in \code{resampling}.}

\item{dml_procedure}{Double machine learning algorithm to be used, either \code{"dml1"} or \code{"dml2"} (default).}

\item{inf_model}{Inference model for final estimation, default \code{"IV-type"} (...)}

\item{se_type}{Method to estimate standard errors. Default \code{"ls"} to estimate usual standard error from least squares regression of residuals. Alternatively, specify \code{"IV-type"} or \code{"DML2018"} to obtain standard errors that correspond to the specified \code{inf_model}. The options chosen for \code{inf_model} and \code{se_type} are required to match.}

\item{bootstrap}{Choice for implementation of multiplier bootstrap, can be set to \code{"normal"} (by default), \code{"none"}, \code{"Bayes"}, \code{"wild"}.}

\item{nRep}{Number of repetitions for multiplier bootstrap, by default \code{nRep=500}.}

\item{...}{further options passed to underlying functions.}
}
\value{
Result object with estimated coefficient and standard errors.
}
\description{
Double Machine Learning for Partially Linear Regression.
}
