% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{print.dca}
\alias{print.dca}
\title{Print an dca object}
\usage{
\method{print}{dca}(x, ...)
}
\arguments{
\item{x}{The result from the dca function.}

\item{...}{ignored}
}
\description{
Print an dca object
}
