% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTask.R
\name{print.DML}
\alias{print.DML}
\title{Methods for Inference Task}
\usage{
\method{print}{DML}(x, ...)
}
\arguments{
\item{x}{Object of class \code{DML}.}

\item{...}{arguments passed to print function and other methods.}
}
\description{
Methods for S3 class \code{DML}
}
