% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmd.r
\name{print.GdmDiag}
\alias{print.GdmDiag}
\title{Print an GdmDiag object}
\usage{
\method{print}{GdmDiag}(x, ...)
}
\arguments{
\item{x}{The result from GdmDiag function.}

\item{...}{ignored}
}
\description{
Print an GdmDiag object
}
